// stamina/model/options/StaminaModelOptions.ts
/*
 * Copyright 2016- Game Server Services, Inc. or its affiliates. All Rights
 * Reserved.
 *
 * Licensed under the Apache License, Version 2.0 (the "License").
 * You may not use this file except in compliance with the License.
 * A copy of the License is located at
 *
 *  http://www.apache.org/licenses/LICENSE-2.0
 *
 * or in the "license" file accompanying this file. This file is distributed
 * on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
 * express or implied. See the License for the specific language governing
 * permissions and limitations under the License.
 */
import MaxStaminaTable from "../MaxStaminaTable";
import RecoverIntervalTable from "../RecoverIntervalTable";
import RecoverValueTable from "../RecoverValueTable";

export interface StaminaModelOptions {
    metadata?: string|null;
    maxCapacity?: number|null;
    maxStaminaTable?: MaxStaminaTable|null;
    recoverIntervalTable?: RecoverIntervalTable|null;
    recoverValueTable?: RecoverValueTable|null;
}